delete old code for Search and Filter
