+++ client/src/components/SectionSwipper&Filter&Search/SectionSwipper&Filter&Search.js
@@ -1,8 +1,8 @@
 import { Box, Container } from '@mui/material';
 import React from 'react';
 import SwiperBanner from '../SwiperBanner/SwiperBanner';
-import Filter from '../Filter/Filter copy 2';
-import Search from '../Search/Search copy';
+import Filter from '../Filter/Filter';
+import Search from '../Search/Search';
 import { stylesGridConteiner, stylesSwipper, stylesSearch, stylesFilter } from './styles';
 
 const SectionSwipperFilterSearch = () => {
